Fix condition flag > check, move to separate namespace
